feature/write : [Fix] write Method에서 음수값이 들어간 경우 stoi exception 이 발생하는 case test 코드 추가및 로직 수정

+++ SSD/SSD/write.cpp
@@ -76,7 +76,7 @@ private:
 		if (cmd_arg.at(2)[0] != '0' || cmd_arg.at(2)[1] != 'x') return false;
 		if (is_valid_hex_string(cmd_arg.at(2).substr(2)) == false) return false;
 		try {
-			write_data = std::stoi(cmd_arg.at(2), 0, 16);
+			write_data = std::stoul(cmd_arg.at(2), 0, 16);
 		}
 		catch (std::exception e) {
 			return false;
